fix(schedule): split schedule draft helpers out of the editor component
CI's Web job runs `npm run lint` as well as build, and lint failed:
ScheduleEditor.tsx exported a component alongside draftFromSchedule,
bodyFromDraft and validateDraft, which trips
react-refresh/only-export-components (3 errors). Build and typecheck both
passed, which is why this was missed locally.

Moved the draft types and helpers to scheduleDraft.ts so the component
file exports only a component. This follows the repo's existing
convention — components/setup/shared.tsx, components/setup/styles.ts and
monitor/chat/styles.ts are all split the same way, and there is not a
single eslint-disable for this rule anywhere in the tree — rather than
suppressing the warning.

Pure move: no behaviour change, and validateDraft's body is byte-identical
to what was removed. Verified with `npm run lint` (exit 0) and
`npm run build` (exit 0); the one remaining warning is the pre-existing
TTSSection.tsx exhaustive-deps one that is also present on main.

Note the OTHER failing CI check, Python (hal) test_startup_volume_parsed,
is NOT from this branch: robots/lamp/ROBOT.md on main sets
startup_volume 40 while hal/test/test_device.py asserts 65. That
contradiction was introduced on main by dc669aec and this branch touches
no files under hal/ or robots/.
